src/main/java/org/sainsbury/controller/ProductController.java
package org.sainsbury.controller;

import lombok.RequiredArgsConstructor;
import org.sainsbury.component.ResponseBuilder;
import org.sainsbury.service.ProductService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/product")
@RequiredArgsConstructor(onConstructor = @__(@Autowired))
public class ProductController {

    private final ProductService productService;
    private final ResponseBuilder responseBuilder;

    @GetMapping(value = "/all", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
    public ResponseEntity getAll() {
        return responseBuilder.success(productService.getAllProducts());
    }
}
